requirements.txt
asgiref==3.4.1
certifi==2021.10.8
chardet==3.0.4
click==8.0.3
decorator==5.1.0
defusedxml==0.7.1
dictdiffer==0.9.0
dj-rest-auth
Django==3.2.9
django-allauth==0.42.0
django-cors-headers==3.10.0
django-environ==0.8.1
django-rest-auth
djangorestframework==3.12.4
djongo==1.3.6
environ==1.0
idna==2.10
Jinja2==3.0.3
jsonpath-rw==1.4.0
MarkupSafe==2.0.1
mongoengine==0.23.1
mongoengine-migrate==0.0.2
oauthlib==3.1.0
ply==3.11
PyJWT==2.3.0
pymongo==3.12.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.24.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.2.4
urllib3==1.25.11
wrapt==1.13.3
